+++ src/core/column-type/image.column.d.ts
@@ -3,8 +3,8 @@ import { DataColumnModel } from './data.column.model';
 import { ColumnModel } from './column.model';
 
 export declare class ImageColumnModel extends DataColumnModel {
-	canUpload?: boolean;
-	hasPreview?: boolean;
+	canUpload?: () => boolean;
+	hasPreview?: (name: string) => boolean;
 }
 
 export declare class ImageColumn extends ColumnView {